Move sign tx functions to utils
